Made a slew of changes.
Changed the Reference Layers category to be a GroupLayer, with each sublayer being a unique reference layer.

Changed the basemap from dark-gray to streets-vector so as to provide a less gloomy look and less contrast with the Street base layer.

Changed the initial coordinates of focus as well as the zoom level to give a better overall view of Los Angeles County.

Eliminated the redundant Esri World Geocoder and LA County CAMS SearchSources.

Implemented a change that will append the query string of a search query to the URL. Still would like to work on removing the query string from the URL when the user no longer wishes to view the result of a query.

Added a Home button widget that will return the user to the original coordinates and zoom level.

Implemented a band-aid solution to eliminate the margins and unused space around the Esri map component. May have to clean up the code a bit more for this change.

Threw in a CoordinateConversion widget just for fun. If the user stories I come up with in a future sprint planning don't involve the use of this widget, I will remove it.