Made another slew of changes.
Moved the initializations of all layers (base layers and reference layers) to the Layers.tsx file to make the code modularized and easier to understand.

Changed the parcel highlight from yellow to watermelon red to better reflect the color used for the highlights in the production website.

Added a Locate widget that makes the map view zoom in on the user's current location.

Threw the Search widget into an Expand widget.

Fixed and configured the tooltips and icons for the Expand widgets.

Added a ScaleBar widget that indicates distances on the map.

Added a Legend widget and threw it into its own Expand widget.

Did some more code refactoring/cleanup.