Made a few changes.
The Popup that appears when a parcel is selected has a differently formatted title. The acronym "AIN" appears, and the AIN itself is technically the APN, the Assessor Parcel Number, which is exactly the same as the AIN, but hyphenated (e.g. 1234-567-890).

Limited the search results to LA County addresses only.

Changed/Updated/Added layers.