Final slew of required changes.
Rolled back some updates to the URLs for the satellite imagery. Added an ActionButton to the Popups of the Assessor Parcels reference layer. Delayed the reissuing of search queries appended to the URL, until the entire map component is loaded (this prevents a race condition in which the program attempts to issue the query before the relevant data is loaded). Added a New Tab button to the bottom-right corner of the map container. Reorganized the code to facilitate embedding the map component into another application/website as an applet.