Updated the Popup's "Get Parcel Detail".
The button darkens when the mouse hovers over it, brightens back up when the mouse moves away, and has an even darker 3D appearance of being pressed down when the mouse is clicked on it.

There's an additional, temporary change to the widgetToggle conditional statement that controls the appearance of the openNewWindowButton. I'm keeping it visible for testing purposes, then will un-comment the exclamation mark once testing is over.