Made yet another slew of changes.
Modularized the code such that the layer initializations and the layer subsets are separate. Layers are divided into subsets, which can be selected based on use case (e.g. internal, external, etc.).

Overhauled the logic for formatting the text in Popups. Added logic that changes the color of the Get Parcel Detail button of the Popup in response to mouse hover events.

Modified the initialization of the Assessor Parcels layer to use the Tile cache.

Changed the behavior of certain layers such that they appear or disappear in response to other layers appearing or disappearing.

Made the legends of layers with one color key invisible, since it's obvious what a color stands for in a layer with only one color key.

Added more layers.

Created a highlight symbol for the Parcels layer that highlights queried parcels.

Added mouse hover events for the Open New Window button.

(Some changes have been made to the search and URL update logic, but more work needs to be done make the logic refined.)

(To-do: the URL that the Open New Window button visits must match the current map search URL used by the current public assessors portal.)